refactor(RVCV-71): 패키지의 오타를 수정한다
