+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ShopAdminUseCase.java
@@ -1,13 +1,13 @@
-package com.romanticpipe.reviewcanvas.domain.shopadmin.aplication.usecase;
+package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase;
 
 import java.util.List;
 
 import org.springframework.web.multipart.MultipartFile;
 
 import com.romanticpipe.reviewcanvas.domain.ReviewDesign;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.aplication.usecase.request.SignUpRequest;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.aplication.usecase.request.UpdateReviewDesignRequest;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.aplication.usecase.response.GetReviewVisibilityTitleResponse;
+import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.SignUpRequest;
+import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.UpdateReviewDesignRequest;
+import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.response.GetReviewVisibilityTitleResponse;
 
 public interface ShopAdminUseCase {
