+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/request/SignUpRequest.java
@@ -1,6 +1,7 @@
-package com.romanticpipe.reviewcanvas.domain.shopadmin.aplication.usecase.request;
+package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request;
 
 import com.romanticpipe.reviewcanvas.domain.ShopInstallType;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
