+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/response/GetReviewVisibilityTitleResponse.java
@@ -1,9 +1,9 @@
-package com.romanticpipe.reviewcanvas.domain.shopadmin.aplication.usecase.response;
-
-import io.swagger.v3.oas.annotations.media.Schema;
+package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.response;
 
 import java.util.List;
 import java.util.Objects;
+
+import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(name = "GetReviewVisibilityTitleResponse", description = "리뷰 노출 항목 title 조회 응답")
 public record GetReviewVisibilityTitleResponse(
